// src/product/models/Product.ts
export interface Product {
    id: number;
    stock: number;
    nombre: string;
    precio: number;
    description: string;
    formula: string;    
    efectos_secundarios: string;
    caducidad: string;
    created_at: string;
    created_by: string;
    updated_at: string;
    updated_by: string;
    deleted: boolean;
}